BE/requirements.txt
absl-py==2.1.0
accelerate==0.33.0
anyio==4.4.0
argbind==0.3.9
asttokens==2.4.1
audioread==3.0.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cmake==3.30.2
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
descript-audio-codec==1.0.0
descript-audiotools==0.7.2
docstring_parser==0.16
einops==0.8.0
exceptiongroup==1.2.2
executing==2.0.1
fastapi==0.95.1
ffmpy==0.4.0
filelock==3.15.4
fire==0.6.0
flatten-dict==0.4.2
fonttools==4.53.1
fsspec==2024.6.1
future==1.0.0
grpcio==1.65.4
h11==0.14.0
huggingface-hub==0.24.5
idna==3.7
importlib_resources==6.4.2
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
julius==0.2.7
kiwisolver==1.4.5
lazy_loader==0.4
librosa==0.10.2.post1
lit==18.1.8
llvmlite==0.43.0
Markdown==3.6
markdown-it-py==3.0.0
markdown2==2.5.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.8
networkx==3.3
numba==0.60.0
numpy==1.24.2
# nvidia-cublas-cu11==11.10.3.66
# nvidia-cublas-cu12==12.1.3.1
# nvidia-cuda-cupti-cu11==11.7.101
# nvidia-cuda-cupti-cu12==12.1.105
# nvidia-cuda-nvrtc-cu11==11.7.99
# nvidia-cuda-nvrtc-cu12==12.1.105
# nvidia-cuda-runtime-cu11==11.7.99
# nvidia-cuda-runtime-cu12==12.1.105
# nvidia-cudnn-cu11==8.5.0.96
# nvidia-cudnn-cu12==9.1.0.70
# nvidia-cufft-cu11==10.9.0.58
# nvidia-cufft-cu12==11.0.2.54
# nvidia-curand-cu11==10.2.10.91
# nvidia-curand-cu12==10.3.2.106
# nvidia-cusolver-cu11==11.4.0.1
# nvidia-cusolver-cu12==11.4.5.107
# nvidia-cusparse-cu11==11.7.4.91
# nvidia-cusparse-cu12==12.1.0.106
# nvidia-nccl-cu11==2.14.3
# nvidia-nccl-cu12==2.21.5
# nvidia-nvjitlink-cu12==12.6.20
# nvidia-nvtx-cu11==11.7.91
# nvidia-nvtx-cu12==12.1.105
packaging==24.1
parler_tts @ git+https://github.com/huggingface/parler-tts.git@8e465f1b5fcd223478e07175cb40494d19ffbe17
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
pooch==1.8.2
prompt_toolkit==3.0.47
protobuf==3.19.6
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==1.10.17
pydub==0.25.1
Pygments==2.18.0
pyloudnorm==0.1.1
pyparsing==3.1.2
pystoi==0.4.1
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytorch-triton==0.0.1
PyYAML==6.0.2
randomname==0.2.1
regex==2024.7.24
requests==2.32.3
rich==13.7.1
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.0
sentencepiece==0.2.0
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
soxr==0.4.0
stack-data==0.6.3
starlette==0.26.1
sympy==1.13.1
tensorboard==2.17.1
tensorboard-data-server==0.7.2
termcolor==2.4.0
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.5.0.dev20240815+cu121
torch-stoi==0.2.1
torchaudio==2.4.0.dev20240815+cu121
torchvision==0.20.0.dev20240815+cu121
tqdm==4.66.5
traitlets==5.14.3
transformers==4.43.3
triton==3.0.0
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.22.0
wcwidth==0.2.13
Werkzeug==3.0.3
TTS==0.22.0
python-dotenv==1.0.1
azure-cognitiveservices-speech
